LCORE-641: split unit tests for HTTP responses models
